AP-4400: Move opponent related "integration" tests to unit specs
Unit test the OtherPartyAttributeGenerator in isolation.
