+++ spec/services/ccms/requestors/case_add_requestor_spec.rb
@@ -262,87 +262,6 @@ module CCMS
               .and have_xml("#{person_xpath}/casebio:RelationToCase", "OPP")
           end
         end
-
-        context "with an existing opponent organisation" do
-          let(:local_authority) do
-            create(:opponent,
-                   :for_organisation,
-                   organisation_name: "Bucks Council",
-                   organisation_ccms_type_code: "LA",
-                   organisation_ccms_type_text: "Local Authority",
-                   ccms_opponent_id: "222222",
-                   exists_in_ccms: true)
-          end
-
-          let(:opponents) { [local_authority] }
-
-          it "generates expected xml" do
-            other_party_xpath = "//casebio:OtherParties/casebio:OtherParty"
-            organization_xpath = "#{other_party_xpath}/casebio:OtherPartyDetail/casebio:Organization"
-
-            expect(request_xml)
-              .to have_xml("#{other_party_xpath}/casebio:OtherPartyID", "222222")
-              .and have_xml("#{other_party_xpath}/casebio:SharedInd", "true")
-              .and have_xml("#{organization_xpath}/casebio:OrganizationName", "Bucks Council")
-              .and have_xml("#{organization_xpath}/casebio:OrganizationType", "LA")
-              .and have_xml("#{organization_xpath}/casebio:RelationToClient", "NONE")
-              .and have_xml("#{organization_xpath}/casebio:RelationToCase", "OPP")
-          end
-        end
-
-        context "with a new opponent organisation" do
-          let(:local_authority) do
-            create(:opponent,
-                   :for_organisation,
-                   organisation_name: "Foobar Council",
-                   organisation_ccms_type_code: "LA",
-                   organisation_ccms_type_text: "Local Authority",
-                   ccms_opponent_id: "")
-          end
-
-          let(:opponents) { [local_authority] }
-
-          before do
-            allow(CCMS::OpponentId).to receive(:next_serial_id).and_return(999_888)
-          end
-
-          it "generates expected xml" do
-            other_party_xpath = "//casebio:OtherParties/casebio:OtherParty"
-            organization_xpath = "#{other_party_xpath}/casebio:OtherPartyDetail/casebio:Organization"
-
-            expect(request_xml)
-              .to have_xml("#{other_party_xpath}/casebio:OtherPartyID", "OPPONENT_999888")
-              .and have_xml("#{other_party_xpath}/casebio:SharedInd", "false")
-              .and have_xml("#{organization_xpath}/casebio:OrganizationName", "Foobar Council")
-              .and have_xml("#{organization_xpath}/casebio:OrganizationType", "LA")
-              .and have_xml("#{organization_xpath}/casebio:RelationToClient", "NONE")
-              .and have_xml("#{organization_xpath}/casebio:RelationToCase", "OPP")
-          end
-        end
-
-        context "with an involved child" do
-          let(:involved_child) { create(:involved_child, legal_aid_application:, full_name: "Billy Elliot", date_of_birth: Date.new(2020, 1, 1)) }
-
-          before do
-            allow(CCMS::OpponentId).to receive(:next_serial_id).and_return(777_777)
-            legal_aid_application.involved_children = [involved_child]
-          end
-
-          it "generates expected xml" do
-            other_party_xpath = "//casebio:OtherParties/casebio:OtherParty"
-            person_xpath = "#{other_party_xpath}/casebio:OtherPartyDetail/casebio:Person"
-
-            expect(request_xml)
-              .to have_xml("#{other_party_xpath}/casebio:OtherPartyID", "OPPONENT_777777")
-              .and have_xml("#{other_party_xpath}/casebio:SharedInd", "false")
-              .and have_xml("#{person_xpath}/casebio:Name/common:Title", "")
-              .and have_xml("#{person_xpath}/casebio:Name/common:FirstName", "Billy")
-              .and have_xml("#{person_xpath}/casebio:Name/common:Surname", "Elliot")
-              .and have_xml("#{person_xpath}/casebio:DateOfBirth", "2020-01-01")
-              .and have_xml("#{person_xpath}/casebio:RelationToClient", "UNKNOWN")
-              .and have_xml("#{person_xpath}/casebio:RelationToCase", "CHILD")
-          end
-        end
       end
     end
   end
